refactor(index): Split main index for API and CLI usage
- Separated the main index file to create a distinct entry point for API-related functionality.
- Introduced a new module called `bin/cli.js` specifically designed to handle command-line interface (CLI) operations.
- Moved command line arguments parser functionality to a module called `argparser` in "bin" directory.
